[TASK] Simplify and clean up error log system information
Remove inheritance from abstract log module

Inheritance was used only to be able to get the time
from which to start to look for errors. This however never
really worked, because the constraint only had a startTimestamp
value set, if the user selected a custom time frame when viewing the log.

Therefore we can just assume errors have not been seen, unless
they are confirmed through the system information toolbar.

Resolves: #85517
Releases: master
Change-Id: Id6aa21fbbe0c56baeb19633aba0f9f7ff6215c0e
Reviewed-on: https://review.typo3.org/57513
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Helmut Hummel <typo3@helhum.io>
Reviewed-by: Helmut Hummel <typo3@helhum.io>