[TASK] Avoid using ControllerContext, and use Extbase Request instead
This change adds the methods
* setRequest()
* getRequest()
* getUriBuilder()

to the RenderingContext of EXT:Fluid.

The main goal is to reduce the usages of
the ControllerContext as much as possible to
decouple Extbase from Fluid.

When the "setRequest" method is used in the renderingContext,
the controllerContext is filled as well, in order to be
backwards-compatible.

Resolves: #92826
Releases: master
Change-Id: I41b8741e947c78895317ef2235959ceb251e103c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66323
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>