[TASK] Set FlashMessageQueue for ModuleTemplate in extbase controller
For the Module Template Layout to render FlashMessages, it
needs to hold the FlashMessageQueue with the added messages.
This needs to be done manually for extbase extensions, as
they are holding their own internal queue.

This patch also reverts the related patch #97569, which
fixed it by adding an explicit f:flashMessages VH to the
templates.

Resolves: #97717
Related: #97569
Related: #96745
Releases: main
Change-Id: I926298f303a4d0d0ffdaffa771ac7667f971b8f1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74795
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>