[BUGFIX] Use proper query for specific belog "Delete similar errors"
In case a `sys_log` entry is deleted via the `belog` backend module,
an error was thrown. This only happened for log entries with
`details_nr` 11 or 12 (PHP errors) AND where placeholder values in
the column `details` were used.

Then the backend log module would try to delete all rows only
with the exactly matching, substituted log details, so that not
too many records would be deleted (with different placeholders).

The code for this tried to unserialize the log data, which led
to the created delete-query utilizing an array as named parameter
placeholder value, which then threw an error.

The solution is to use the actual raw data for the query comparison,
which is what this patch does. Then the exact contents of the
fields `details` and `log_data` will be used to delete all
matching rows.

Resolves: #104494
Releases: main, 12.4

Change-Id: I393b918919340eb279c62136feec2f27eeac708f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85402
Reviewed-by: Guido Schmechel <guido.schmechel@brandung.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>