[!!!][FEATURE] Support qualified and unqualified ISO8601 in DataHandler
Streamline DataHandler datetime field processing to support unqualified
ISO8601 dates instead of misusing fake UTC-0 ISO8601 for representing
LOCALTIME [1][2][3].

This allows DataHandler to correctly interpret offset information (like
a real UTC-0 datestring) and to perform proper transformation into the
respective database representation. (Original timezone information is
still normalized to UTC (integer fields) or localtime (native datetime
fields), but the point-in-time is correctly persisted.)

The TYPO3 backend date picker element is adapted to use unqualified
ISO8601 LOCALTIME as interchange format between FormEngine PHP and
HTML/JS code.

Why use localtime without offset information (=unqualified)?
------------------------------------------------------------

Conceptually this is not a change. It is a technical cleanup of the
current approach as we have been effectively using ISO8601 localtime
as interchange format between FormEngine and DataHandler before as well:

A faked ISO8601 UTC0 string was used that was interpreted as
localtime. For example 1999-11-11T11:11:11Z on a server timezone
Europe/Berlin was interpreted as:
1999-11-11T11:11:11+01:00
           ^^

Formally seen that was wrong as 1999-11-11T11:11:11Z actually refers to
the 12th hour in Germany, when taking the time zone database rules for
Europe/Berlin in November 1999 (=CET with offset +01:00) into account:
1999-11-11T12:11:11+01:00
           ^^

That means we actually misused UTC offset Z(ero) als localtime qualifier,
which is wrong because it refers to UTC+0 and thus formally to another
point in time that we actually tried to encode.

As a consequence we had to subtract the servers current timezone offset
after parsing the fake UTC0 date string leading to complex code and
misunderstandings which introduced bugs like #103693.

Using explicit localtime basically allows to remove all these
workarounds and results in much simpler code.

Why not switch to real UTC + offsets directly?
----------------------------------------------

Editors (currently) do only enter dates, no timezone offsets or
timezone names.
They are supposed to enter dates in the servers timezone (which is
currently not introspectable from the backend UI, but that's another
topic and a UX issue).

To transform the raw user input into the server timezone it's important
to know what the *user intended* date and time is, that means we need to
know the date and time values the user has selected and not a mapped
representation in UTC, otherwise the JS datepicker code would need to be
aware of the server timezone including timezone rules.

That's possible but actually a major task. A simple "current" timezone
offset (like +01:00 for Europe/Berlin during winter time) would not be
sufficient, since the datepicker would need to be aware of future
daylight saving times and adjust offsets in order to create a correct
UTC-offset for the selected time to match the server timezone (for
example when picking a datetime from the summer time range while it's
currently winter time).

Therefore, as long as we do not expose a timezone selection to the
FormEngine frontend, we strive for the simplicity of raw user input and
perform the transformation into a date + timezone on the server.

This approach has also been chosen to be used by WHATWG[3][4] for
the `<input type="datetime-local">` element, where qualified
ISO8601 dates are explicitly forbidden, instead an unqualified
ISO8601 localtime date is expected.

Adopting to this paradigm will allow us to replace our custom datetime
picker with the native `<input type="datetime-local">` at some point.
(Currently we're not happy with the Firefox UX and therefore stay with
a custom JavaScript based date picker, but it's good to be prepared to
switch at any point in time by using a compatible interchange format).

That means: Switching to localtime now will pave the way for timezones
in UI later on, because we're than able to honor custom timezone
information as described in the example in [3] (airpoirt datetime-local
timezone <select> field).

API changes
-----------

Now that we adapt DataHandler to accept localtime for FormEngine to
DataHandler communication, we also officially support datetime offsets
to be provided for DataHandler-as-an-API usage, without the previously
quite "unexpected" result that timezone offsets were cropped,
because of our localtime misinterpretation.

That means, DataHandler now accepts five different *input* formats,
regardless of the persistence method of the respective field
(e.g. native datetime or timestamp field):

   * ISO8601 in LOCALTIME (Y-m-d\\TH:i:s):
     1999-11-11T11:11:11
     (interchange format between TYPO3 backend server/html/js)

   * ISO8601 with timezone offsets (Y-m-d\\TH:i:sP):
     1999-11-11T10:11:11Z
     1999-11-11T11:11:11+01:00
     (To be used as API)

   * SQL flavored date/times (Y-m-d H:i:s):
     1999-11-11 11:11:11
     1999-11-11
     11:11:11
     (To be used for copy operations of native datetime fields)

   * Unix timestamps (U):
     942315071
     (To be used for copy operations of unix timestamp fields)

   * DateTime objects with timezone offsets.
     NEW! Allows to pass-in `\DateTimeInterface` objects as-is without
     requiring a to-string conversion.
     // same effect as ISO8601 LOCALTIME above
     new \DateTimeImmutable('1999-11-11T11:11:11')
     // same effect as ISO8601 with offset above
     new \DateTimeImmutable('1999-11-11T11:11:11+00:00')
     new \DateTimeImmutable('1999-11-11T12:11:11+01:00')
     // Offsets of custom timezones will be accepted an converted to
     // UTC for timestamp persistence and
     // LOCALTIME for native datetime fields
     // Heads UP: That means timezone offset is lost on persistence,
     //           but the date in time is stored correctly, without
     //           cropping offsets
     new \DateTimeImmutable(
         '1999-11-11T11:11:11',
         new \DateTimeZone('Asia/Jerusalem')
     )

This change is "breaking" because the previous fake ISO8601 UTC0 strings
will now be interpreted as *real* UTC0 and the offset to the server
timezone will be normalized instead of being cropped.

That means: For those who used DataHandler as an API, adaptions will be
needed to remove previous "workarounds" for DataHandlers broken DateTime
handling.

[1] https://en.wikipedia.org/wiki/ISO_8601#Local_time_(unqualified)
[2] https://html.spec.whatwg.org/multipage/common-microsyntaxes.html#local-dates-and-times
[3] https://html.spec.whatwg.org/multipage/input.html#local-date-and-time-state-(type=datetime-local)
[4] https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input/datetime-local

Releases: main
Resolves: #105549
Related: #104309
Change-Id: Iaaf823ee57564e8835ac129707def927ae49804f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86888
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>